ipacm: Changes to handle calculation overflow
Added changes to handle IPACM file size calculation overflow while
calculating the max file size.

+++ ipacm/inc/IPACM_Xml.h
@@ -26,9 +26,9 @@
  * OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN
  * IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  *
- * Changes from Qualcomm Innovation Center, Inc. are provided under the following license:
- * Copyright (c) 2025 Qualcomm Innovation Center, Inc. All rights reserved.
- * SPDX-License-Identifier: BSD-3-Clause-Clear.
+ * Changes from Qualcomm Technologies, Inc. are provided under the following license:
+ * Copyright (c) Qualcomm Technologies, Inc. and/or its subsidiaries.
+ * SPDX-License-Identifier: BSD-3-Clause-Clear
  */
 /*!
   @file
@@ -347,7 +347,7 @@ typedef struct  _IPACM_conf_t
 	bool lan_stats_enable;
 #endif
 
-	uint32_t max_file_size;
+	int64_t max_file_size;
 	uint8_t max_file_size_quota;
 	bool ipv6_nat_enable;
 	int ipacm_l2tp_enable;
